Move `editorAssociations` to own file
